release: protocol v2 end-to-end pipeline, CLI toolchain, auto-tuned PID
Firmware (STM32F429):
- PROTOCOL v2: variable-length frames (0x55AA/LEN/CRC8/SEQ) with ECG,
  SpO2, IBP/NIBP, temperature and accessory-status message IDs
  (3.FIRMWARE/protocol.h + PROTOCOL.md)
- physio_app application layer with SIM/real dual mode (PHYSIO_SIM_MODE);
  AFE4490 pulse-oximeter driver ported from the verified LPC177x project
- ADS1294 -> ADS1298R rename; legacy 118-byte frame path removed
- Auto-tuned heater PID (FF + PI + D-on-measurement, anti-windup,
  bleed-on-zero-cross, slew limit): setpoint steps settle in 95-155s
  with <=0.01C overshoot; cold start reaches 38C with zero overshoot
- UART5 @ 819200 baud (exact divider on the 32.768MHz APB clock)

ESP32 bridge:
- Fixed uart_rx_task stack overflow that corrupted the heap and caused
  a boot loop (buffer made static, stack 6KB)
- UART pins corrected per board netlist (IO5/IO18), real-data mode default

PC console (4.Python_console, dir renamed):
- Length-aware protocol parser with 7 frame types, SpO2/temperature
  status bar, 30-minute rolling temperature trend plot
- float32 plot path fixes (pyqtgraph overflow warnings, missing import)

Tooling:
- build.py / flash.py: zero-dependency command-line build & flash
  (J-Link preferred, OpenOCD fallback) + TOOLCHAIN.md / FLASH_GUIDE.md

Comments in new code translated to English; test/tuning artifacts
excluded from the repository.

+++ 4.Python_console/V2/ui/waveform_widget.py
@@ -123,7 +123,8 @@ class WaveformWidget(QWidget):
             return
 
         # 添加 4 个样本到缓冲区 (不是 10 个)
-        self.buffer.add_data(frame.samples)
+        # float32: 避免 pyqtgraph 对 ±400万原始计数做 int 运算时溢出告警
+        self.buffer.add_data(frame.samples.astype(np.float32))
 
         # 更新 Y 轴范围
         self._update_y_ranges()
